requirements.txt
tk
ttkthemes
qrcode
pillow
